backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
Pyro5>=5.14
pydantic>=2.0
supabase>=2.0.0
python-dotenv>=1.0.0
